# design/api/openapi.yml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  version: '1.0'
  description: 'REST API to be used to create tableside food ordering applications.'
  contact:
    email: mawentowski@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:4010/
    description: Prism proxy server
  - url: http://localhost:8080/
    description: API server
...
paths:
...
  /categories:
    post:
      summary: Create a category
      description: |
        This operation creates a new category with the provided information. Categories can be used to organize/group dishes in food ordering apps. For example, displaying tabs for each category. 

        The `category_id` returned in the response can be used to associate the category with a menu.
      operationId: postCategory
      tags:
        - Categories
      security:
        - AccessToken: []
      requestBody:
        $ref: '#/components/requestBodies/CreateCategoryRequest'
      responses:
        '201':
          description: |
            Resource created successfully. The response body contains the newly created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          headers:
            Location:
              schema:
                type: string
              description: Points to the resource endpoint for the newly created resource
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    get:
      summary: Get a list of categories
      description: This operation retrieves a list of categories based on the specified criteria.
      operationId: getCategoryList
      tags:
        - Categories
      parameters:
        - name: sort
          in: query
          description: |
            Sort fields in ascending or descending order. To sort ascending, postfix the field name with `:asc`. To sort descending, postfix the field name with `:desc`.
          schema:
            type: string
          example: 'name:asc'
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,dish_ids'
        - name: select
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator:value`.

            **Basic Equality (eq):**
            - For exact matches, use `eq:value`.

            **Comparison Operators:**
            - **Less Than (lt):** Use `lt:value`.
            - **Less Than or Equal To (lte):** Use `lte:value`.
            - **Greater Than (gt):** Use `gt:value`.
            - **Greater Than or Equal To (gte):** Use `gte:value`.

            Separate multiple filters with commas. Dates should be in `YYYY-MM-DD` format.
          schema:
            type: string
            example: 'created_at.gt:2024-04-23,name.eq:main_course'
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response. If `limit` is used in conjunction with `offset`, then page-based pagination is applied.
          schema:
            type: integer
            maximum: 30
          example: 20
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources. A value of `0` will display resources from the beginning of the list. Only affects the response data if a `limit` value is set.
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  total_results:
                    type: integer
                    example: 1

        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /categories/{category_id}:
    get:
      summary: Get a category's details
      description: This operation retrieves information about a category identified by its unique category ID.
      operationId: getCategory
      tags:
        - Categories
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
          example: '632a14b6d004f3001f12c8c8'
          description: The unique identifier for the resource.
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
          example: 'name,dish_ids'
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      summary: Update a category
      description: This operation updates an existing category with the provided data.
      operationId: putCategory
      tags:
        - Categories
      security:
        - AccessToken: []
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
          example: '643e8ef8b5a1e8001f7ecb16'
          description: The unique identifier for the resource.
      requestBody:
        $ref: '#/components/requestBodies/UpdateCategoryRequest'
      responses:
        '200':
          description: Successful operation. The response body contains the updated resource.
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request could not be completed due to a conflict with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      summary: Delete a category
      description: This operation deletes an existing category identified by its unique category ID.
      operationId: deleteCategory
      tags:
        - Categories
      security:
        - AccessToken: []
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
          example: '63c896f844697e001f1a19c0'
          description: The unique identifier for the resource.
      responses:
        '204':
          description: No content. The operation was successful, and there is no content to return.
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /categories/{category_id}/dishes:
    get:
      summary: View a category's dishes
      description: This operation retrieves a list of dishes belonging to a specific category. Each dish object returned contains detailed information about the dish and which ingredients are essential to preparing the dish.
      operationId: getCategoryDishes
      tags:
        - Categories
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
          example: '65bf24fc1f1b5c001f1d99fd'
          description: The unique identifier for the resource.
        - name: include
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          schema:
            type: string
      responses:
        '200':
          description: Successful operation. The response body contains the requested data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dish'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description: Unauthorized. Authentication is required, or the provided credentials are invalid.
        '403':
          description: Forbidden. The server understood the request, but it refuses to authorize it. Authentication will not help.
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error. An unexpected error occurred on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  requestBodies:
    CreateCategoryRequest:
      description: Request body containing information used to create the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              dish_ids:
                $ref: '#/components/schemas/DishIds'
              name:
                $ref: '#/components/schemas/Name'
            required:
              - dish_ids
              - name
    UpdateCategoryRequest:
      description: Request body containing information used to update the resource.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              dish_ids:
                $ref: '#/components/schemas/DishIds'
              name:
                $ref: '#/components/schemas/Name'
...
  schemas:
...
    Category:
      type: object
      properties:
        category_id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        dish_ids:
          $ref: '#/components/schemas/DishIds'
        name:
          $ref: '#/components/schemas/Name'
...
    CreatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the resource was created in ISO 8601 format.
    Description:
      type: string
      example: A description of the resource.
    Dish:
      type: object
      properties:
        dish_id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        description:
          $ref: '#/components/schemas/Description'
        image_name:
          $ref: '#/components/schemas/ImageName'
        ingredients:
          $ref: '#/components/schemas/Ingredients'
        name:
          $ref: '#/components/schemas/Name'
        price:
          $ref: '#/components/schemas/Price'
        station:
          $ref: '#/components/schemas/Station'
    DishIds:
      type: array
      items:
        type: string
      example:
        - '692a4a34a46ecf001f35ac6f'
        - '6a4a857c022eb7001f9b8972'
        - '6b7c2441a0db43001f9bf377'
      description: Contains the unique identifiers of dishes to add to the order.
...
    Id:
      type: string
      example: '674d0bf5c28b69001f8e03a1'
      readOnly: true
      description: The unique identifier for the resource.
    ImageName:
      type: string
      description: |
        API consumers can use this string to generate the asset URL for displaying the dish image in user interfaces.
      example: 'burger.jpg'
    Ingredients:
      type: array
      description: List of ingredients used during the preparation of the dish.
      items:
        type: object
        properties:
          ingredient_id:
            type: string
            example: '66294b2a4475a41f3e709bc4'
            description: The unique identifier for the resource.
          is_essential:
            type: boolean
            description: |
              Whether the ingredient is essential to prepare the dish.
            example: true
        example:
          - ingredient_id: '66294b2a4475a41f3e709bc5'
            is_essential: true
          - ingredient_id: '66294b2a4475a41f3e709bc6'
            is_essential: true
          - ingredient_id: '66294b2a4475a41f3e709bc7'
            is_essential: false
...
    Name:
      type: string
      example: 'main_course'
      description: ''
...
    Price:
      type: number
      format: float
      example: 9.99
      description: The cost of the dish.
...
    Problem:
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
      required:
        - title
        - detail
...
    Station:
      type: string
      description: |
        The restaurant kitchen has 3 stations where food is prepared:
          - cold: This area is responsible for preparing cold appetizers, salads, and desserts that don't require cooking.
          - hot: Also known as the "hot line" or "hot side," this area handles dishes that require cooking over heat, such as sautéing, frying, grilling, and baking.
          - beverages: In addition to coffee bars, there may be stations dedicated to preparing other beverages such as cocktails, smoothies, or fresh juices.
      enum: ['cold', 'hot', 'beverages']
      example: 'hot'
...
    UpdatedAt:
      type: string
      format: date-time
      example: '2024-04-22T10:00:00Z'
      readOnly: true
      description: Timestamp when the order was created in ISO 8601 format.
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.